// server/enums/otherQualificationsValue.ts
enum OtherQualificationsValue {
  ADVANCED_EDUCATION = 'ADVANCED_EDUCATION',
  CSCS = 'CSCS',
  DRIVING_LICENSE = 'DRIVING_LICENSE',
  FIRST_AID = 'FIRST_AID',
  FOOD_HYGIENE = 'FOOD_HYGIENE',
  HEALTH_AND_SAFETY = 'HEALTH_AND_SAFETY',
  HGV_LICENSE = 'HGV_LICENSE',
  MACHINERY = 'MACHINERY',
  MANUAL = 'MANUAL',
  TRADE = 'TRADE',
  OTHER = 'OTHER',
  NONE = 'NONE',
}

export default OtherQualificationsValue